requirements.txt
flask_sqlalchemy
flask_migrate
flask_script
flask_mail
flask_user
tweepy
